+++ examples/firmware_bundles.py
@@ -34,8 +34,10 @@ config = {
     }
 }
 
-# To run this example you must define a path to a valid file
-firmware_path = "<path_to_firmware_bundle>"
+# Example updates - change to what you want to upload
+spp_path = "SPPgen9snap6.2015_0405.81.iso"
+patch_path = "CP036110.zip"
+compsig_path = "CP036110.compsig"
 
 # Try load config from a file (if there is a config file)
 config = try_load_from_file(config)
@@ -43,6 +45,18 @@ oneview_client = OneViewClient(config)
 
 # Upload a firmware bundle
 print("\nUpload a firmware bundle")
-firmware_bundle_information = oneview_client.firmware_bundles.upload(file_path=firmware_path)
+
+# Upload an entire SPP
+(bundle_information, compsig_information) = oneview_client.firmware_bundles.upload(file_path=spp_path)
+
 print("\n Upload successful! Firmware information returned: \n")
-pprint(firmware_bundle_information)
+pprint(bundle_information)
+
+# Upload just a single patch
+(bundle_information, compsig_information) = oneview_client.firmware_bundles.upload(
+    file_path=patch_path, compsig_path=compsig_path)
+
+print("\n Upload successful! Firmware information returned: \n")
+pprint(bundle_information)
+print("\n Firmware information returned: \n")
+pprint(compsig_information)